logger outbuf not initialized when ostream ctor is called
A crash happens on the static initialization of eoLogger eo::log;
at the end of the file eoLogger.cpp (line 255)

The reason is that the ctor used to initialize the std::eostream base class of
eoLogger is invoked with an _obuf argument while _obuf has not been initialized
already.

The solution is to call the function
std::ostream::init(&_obuf)
after the ctor initalizer list, i.e. inside the body ctor.

Former-commit-id: d706968c31b04d3c45b6faaf384f544f13aba7e4